Extract modules from main.js
